chore(refactor): move file getters into FileService
