+++ lib/Controller/PublicSessionController.php
@@ -11,7 +11,7 @@ namespace OCA\Text\Controller;
 use OCA\Text\Middleware\Attribute\RequireDocumentBaseVersionEtag;
 use OCA\Text\Middleware\Attribute\RequireDocumentSession;
 use OCA\Text\Service\ApiService;
-use OCA\Text\Service\DocumentService;
+use OCA\Text\Service\FileService;
 use OCP\AppFramework\Http;
 use OCP\AppFramework\Http\Attribute\NoAdminRequired;
 use OCP\AppFramework\Http\Attribute\PublicPage;
@@ -39,6 +39,7 @@ class PublicSessionController extends PublicShareController implements ISessionA
 		private ShareManager $shareManager,
 		private ApiService $apiService,
 		private DocumentService $documentService,
+		private FileService $fileService,
 		private IL10N $l10n,
 	) {
 		parent::__construct($appName, $request, $session);
@@ -73,7 +74,7 @@ class PublicSessionController extends PublicShareController implements ISessionA
 	#[NoAdminRequired]
 	#[PublicPage]
 	public function create(string $token, ?string $filePath = null, ?string $baseVersionEtag = null, ?string $guestName = null): DataResponse {
-		$file = $this->documentService->getFileByShareToken($token, $filePath);
+		$file = $this->fileService->getFileByShareToken($token, $filePath);
 		/*
 			* Check if we have proper read access (files drop)
 			* If not then well 404 it is.
@@ -92,7 +93,7 @@ class PublicSessionController extends PublicShareController implements ISessionA
 	#[NoAdminRequired]
 	#[PublicPage]
 	public function close(int $documentId, int $sessionId, string $sessionToken, string $token): DataResponse {
-		$file = $this->documentService->getFileByIdFromShare($documentId, $token);
+		$file = $this->fileService->getFileByIdFromShare($documentId, $token);
 		return $this->apiService->close($documentId, $sessionId, $sessionToken, $file);
 	}
